Move rounded line, circle and frame draw primitives to kas
